# Reduce memory allocations in array handling