@lahma 2022-10-14
This concept could probably be used also more generally, I'm just afraid the amount of changes to track `JsValue` vs `PropertyDescriptor`. Memory benefits would be obvious for cases like `const obj = { "x": 1, "y": 2, "z": 3 }`, no need to have special properties, they are all plain data.