@sebastienros 2022-10-14
Maybe these classes need multiple dedicated implementations (sparse|dense * JsValue|PropertyDescriptor). For instance are all properties JsValue/PropertyDescriptor, or can only a few of them be PD? Then be able to dynamically convert from dense|jsvalue ro sparse|PD. Could be 4 types, maybe through inheritance. Or do you think most common case is dense+jsvalue then it could only be two classes: dense+jsvalue and sparse+PD, with a dynamic conversion from one to the other when the case is detected.