@lahma 2022-10-15
I think I'll try the same approach as here. We already have `PropertyDictionary` which is hybrid dictionary changing from list-based to hash-key-based trying to optimize for small collections. For arrays switching to sparse (dictionary) it really kills perf if it's done "too soon" - like if it's single special property - but I guess it's mostly the test suite that likes to play these tricks..